Add support for multiple writes in a single loop

+++ benchmarks/main.py
@@ -2,9 +2,10 @@ import glob
 import os
 from typing import List
 
-from benchmark import Benchmark
+from benchmark import Benchmark, ComparisonBenchmark
 from compilation_strategy import ClangStrategy, OptClangStrategy, PolytopeStrategy, ClangO3Strategy
-from example_generator import TestGenerator, RandomLinGenerator, SelectedExampleGenerator, RepeatedExampleGenerator
+from example_generator import TestGenerator, RandomLinGenerator, SelectedExampleGenerator, RepeatedExampleGenerator, \
+    TestExampleGenerator
 from execution_strategy import CorrectnessTest, TimeTest, BarChart, LineGraph
 
 
@@ -14,18 +15,24 @@ def main():
     benchmark = Benchmark(
         # TestGenerator(1000),
         # RandomLinGenerator(1000),
-        RepeatedExampleGenerator("./examples/arr_red.c", 1000),
+        SelectedExampleGenerator(2000),
         [OptClangStrategy(), PolytopeStrategy()],
         # [ClangStrategy(), OptClangStrategy(), PolytopeStrategy(), ClangO3Strategy()],
         # [BarChart(iterations=30, compile_names=["Clang+Opt", "Tope", "Clang+Polly"], test_names=test_names,
         #           normalise=True)],
-        [LineGraph(iterations=10, compile_names=["Clang+Opt", "Tope"])],
-        # [CorrectnessTest(),
-        # TimeTest(iterations=50, names=["Clang", "Clang+Opt", "Polytope", "Clang -O3"])],
-        50,
-        True
+        # [LineGraph(iterations=40, compile_names=["Clang+Opt", "Tope"])],
+        [CorrectnessTest()],
+        7,
+        False
     )
     benchmark.run()
+
+    # comparison_benchmark = ComparisonBenchmark(
+    #     [LineGraph(iterations=10, compile_names=["Baseline", "Tope"])],
+    #     3,
+    #     False
+    # )
+    # comparison_benchmark.run()
 
 
 def create(examples: List[str], name="example"):
@@ -36,10 +43,11 @@ def create(examples: List[str], name="example"):
 
 
 def clear():
-    dump_files = glob.glob("./dump/*")
+    dump_files = glob.glob("./dump/*") + glob.glob("./dump/A/*") + glob.glob("./dump/B/*")
     bin_files = glob.glob("./bin/*")
     for f in dump_files + bin_files:
-        os.remove(f)
+        if os.path.isfile(f):
+            os.remove(f)
 
 
 if __name__ == "__main__":
